% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_aphia.R
\name{add_Aphia_ID}
\alias{add_Aphia_ID}
\title{Add WoRMS Aphia Id
Add WoRMS Aphia Id to predator or prey species where missing.}
\usage{
add_Aphia_ID(
  stom,
  Worms_Aphia_ID,
  predator_or_prey = c("predator", "prey")[2],
  stop_if_errror = FALSE,
  delete_first = FALSE,
  verbose = FALSE,
  match_item = c("item", "scientific")[2]
)
}
\arguments{
\item{stom}{Stomach data set of class STOMobs, from e.g. \link{read_exchange_data}).}

\item{Worms_Aphia_ID}{File Name including WoRMS Aphia ID by latin (or other) predator or prey name.file 'WoRM_AphiaID.csv' in the package /inst directory is used as default.}

\item{predator_or_prey}{Add missing Aphia ID for predator | prey.}

\item{stop_if_errror}{Logical for stopping if AphiaID is not found for all.}

\item{delete_first}{Delete all existing WoRMS Aphia Id before adding new}

\item{verbose}{Show information on progress}

\item{match_item}{Variable name in WoRM_AphiaID.csv ("item" or "scientific') used to assign WoRMS Aphia ID.}
}
\value{
Stomach data on on internal format.
}
\description{
Add WoRMS Aphia Id
Add WoRMS Aphia Id to predator or prey species where missing.
}
\examples{
\dontrun{add_Aphia_ID(X)}
}
