% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_strata.R
\name{add_strata}
\alias{add_strata}
\title{Add temporal and spatial strata to STOMobs}
\usage{
add_strata(stom)
}
\arguments{
\item{stom}{Stomach data set of class STOMobs.}
}
\value{
STOMobs object with added strata.
}
\description{
Add temporal and spatial strata as specified in a control object \link{STOMcontrol-class} for collation of stomach data.
}
\examples{
\dontrun{a<-add_strata(a)}
}
