% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMdiet.R
\name{as.data.frame.STOMdiet}
\alias{as.data.frame.STOMdiet}
\title{Convert STOMdiet into a data.frame}
\usage{
\method{as.data.frame}{STOMdiet}(diet)
}
\arguments{
\item{diet}{Diet dataset of class STOMdiet.}
}
\value{
STOMdiet object transformed to a "flat" data.frame.
}
\description{
Convert STOMdiet into a data.frame
}
