% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMobs.R
\name{as.data.frame.STOMobs}
\alias{as.data.frame.STOMobs}
\title{Convert a STOMobs into a data.frame}
\usage{
\method{as.data.frame}{STOMobs}(stom, add_empty = TRUE)
}
\arguments{
\item{stom}{Stomach data set of class STOMobs.}

\item{add_empty}{Insert prey "other" with prey weight (prey_w) =0 for samples with only empty stomachs.}
}
\value{
STOMobs object transformed to a "flat" data frame.
}
\description{
Method to convert an object of class STOMobs into a data frame (and tibble).
Information on sampling and predators are copied to each record of prey information.
}
\examples{
\dontrun{a<-as.data.frame(stom); a}
}
