% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMdiet.R
\name{as_STOMdiet}
\alias{as_STOMdiet}
\title{Convert a data.frame into a STOMdiet object}
\usage{
as_STOMdiet(d, new_pred_var, new_prey_var)
}
\arguments{
\item{d}{Stomach data set on data.frame format. d must have been created from a call to of class as.data.frame().}

\item{new_pred_var}{Variable names to be added to the Predator data set in STOMdiet.
which were added the call to as.data.frame()}

\item{new_prey_var}{Variable names to be added to the prey data set in STOMdiet}
}
\value{
STOMdiet object
}
\description{
Method to convert an object of class data.frame into an object of class STOMdiet.
}
\examples{
\dontrun{a<-as_STOMdiet(df); head(a)}
}
