% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMobs.R
\name{as_STOMobs}
\alias{as_STOMobs}
\title{Convert a data.frame into a STOMobs object}
\usage{
as_STOMobs(d, new_pred_var, new_prey_var)
}
\arguments{
\item{d}{Stomach data set on data.frame format. d must have been created from a call to of class as.data.frame().}

\item{new_pred_var}{Variable names to be added to the Predator data set in STOMobs.
which were added the call to as.data.frame()}

\item{new_prey_var}{Variable names to be added to the prey data set in STOMobs}
}
\value{
STOMobs object
}
\description{
Method to convert an object of class data.frame into an object of class STOMobs.
}
\examples{
\dontrun{a<-as_STOMobs(df); a}
}
