% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correct_regurgitated.R
\name{bias_correct_regurgitated}
\alias{bias_correct_regurgitated}
\title{Bias correct due to regurgitated stomachs}
\usage{
bias_correct_regurgitated(
  stom,
  delete_just_regurgitated = TRUE,
  drop_variables,
  update_n_food_with_n_regur = TRUE
)
}
\arguments{
\item{stom}{stomach contents data of class STOMobs.}

\item{delete_just_regurgitated}{Delete stomachs where all stomachs within a sample_id and predator size class are regurgitated.}

\item{drop_variables}{Names of variables for number of stomachs to be deleted from dataset. This includes combinations of 'n_food', 'n_regur', 'n_skel' and 'n_empty'.
By default the variables 'n_food', 'n_regur', 'n_skel' and 'n_empty' will be deleted.}

\item{update_n_food_with_n_regur}{Update number of stomachs with food (n_food) by the number of regurgitated stomachs (n_regur).}
}
\value{
stomach contents data of class STOMobs, where the stomach contents per sample have been raised as it is assumed
that the regurgitated stomach had the same stomach contents as the one with food (and no regurgitation).
}
\description{
Regurgitated stomachs are from feeding fish, but their actual absolute stomach contents (in the sea) are not known.
Stomach contents from regurgitated stomachs are not analysed or included as data, however the number
of regurgitated stomachs (n_regur) should be noted together with the number of stomachs with skeletal remains only (n_nskel) and the number of empty stomachs (n_empty).
The bias correction assumes that the regurgitated stomachs have had the same stomach contents as the observed feeding fish (n_food).
}
\details{
Ideally each stomach should be classified as: with food, with food but
regurgitated, with skeletal remains only or empty. This information is
included in the presently used exchange format.

\tabular{ll}{
\strong{Variable } \tab \strong{Comments} \cr
n_food  \tab Stomachs with (recently ingested) food\cr
n_regur \tab Stomach with food, but evidence of regurgitation of parts or all the stomach contents\cr
n_skel  \tab Stomachs with practically indigestible remains and the fish is not considered feeding recently\cr
n_empt  \tab No stomach contents \cr
}

Stomach contents from regurgitated (or everted) stomachs should not be
included in the estimation of diet as the proportion regurgitated is
unknown. The number of fish with regurgitated stomachs may however be
recorded ('n_regur'). With the assumption that the
regurgitated stomachs sampled within 'sample_id' have had the same
stomach contents as the feeding (non-regurgitated) fish each observed
prey item weight and number can be corrected by a factor to calculate
the mean stomach contents of a predator within a 'sample_id'.

The bias correction makes sense for stomach that have been pooled during sampling (e.g. the 1981 and 1991 North Sea sampling)
or if several individually processed stomachs are aggregated into a group of stomachs to get the stomach contents for a group of fish (see  \link{aggregate_within_sample)}.

\deqn{factor = \frac{(N_{food} + N_{regur})}{N_{food}*(N_{food} + N_{skel} + N_{regur} + N_{empt})}}
}
\examples{
\dontrun{a <- bias_correct_regurgitated(stom=a,delete_just_regurgitated=TRUE,
          drop_variables=c('n_skel','n_empty', 'n_regur'))}
}
