% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootsMean}
\alias{bootsMean}
\title{Statistics from bootstrap replicates.}
\usage{
bootsMean(
  b,
  min_prey = 0.01,
  pointEst,
  by_prey_size = FALSE,
  n_rep,
  do_Diri = FALSE
)
}
\arguments{
\item{b}{Bootstrap diet data set. List of elements of class STOMdiet.}

\item{min_prey}{prey weight value for missing combination of preys and prey sizes.}

\item{pointEst}{Diet data estimated without bootstrapping (point estimates ogf prey weights)}

\item{by_prey_size}{Logical for calculating mean and variance by prey size. FALSE provides statistics by prey species.}

\item{n_rep}{Selected replicates used. If n_rep is a number, 1:n_rep are used, If n_rep is a vector the provided replica numbers are used in the statistics}

\item{do_Diri}{Logical for performing statistics assuming the replicates are Dirichlet distributed.}
}
\value{
Data set with mean and variance of diets and fit to Dirichlet distribution. Output includes
\tabular{ll}{
\strong{Variable} \tab \strong{Contents} \cr
nboots    \tab Number of bootstrap replicates.\cr
year      \tab Year.\cr
quarter   \tab Quarter of the year.\cr
pred_name \tab Predator name.\cr
pred_size \tab Predator size.\cr
prey      \tab Prey, prey name and size, if used.\cr
prey_name \tab Prey name.\cr
prey_size \tab Prey size.\cr
n_prey_sp \tab Number of "preys" observed.\cr
param     \tab Estimated parameters assuming a Dirichlet distribution of bootstrap replicates.\cr
phi       \tab Estimated precision parameter (sum of param from the Dirichlet distribution).\cr
mu        \tab Estimated mean value from the Dirichlet distribution.\cr
p_value   \tab p-value for Log-likelihood ratio test for the estimated Dirichlet \cr \tab mean vector describe the bootstrap data well.\cr
mean_w    \tab Mean value of prey weights (proportions).\cr
sd_w      \tab Standard deviation of prey weights (proportions).\cr
prey_w    \tab Point estimate of prey weight (proportions). \cr
strata    \tab Strata used for bootstrap.\cr
n_stom    \tab Number of stomachs available before bootstrap ( from the \code{pointEst} data).\cr
n_sample  \tab Number of samples available before bootstrap.\cr
}
}
\description{
Statistics from bootstrap replicates.
}
