% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootstrap_addVar}
\alias{bootstrap_addVar}
\title{Add bootstrap stratification variables.}
\usage{
bootstrap_addVar(s)
}
\arguments{
\item{s}{Stomach data set of class STOMobs.}
}
\value{
Stomach data set of class STOMobs with added variable boots_strata and boots_id.
}
\description{
Add variables for bootstrapping stratification.
}
\examples{
\dontrun{bootstrap_addVar(stom)}
}
