% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootstrap_show}
\alias{bootstrap_show}
\title{Show bootrap stratification.}
\usage{
bootstrap_show(s, show = c("strata", "sample")[1])
}
\arguments{
\item{s}{Stomach data set of class STOMobs.}

\item{show}{Show either number of stomachs per strata or number of stomachs per samples available for bootstrapping.}
}
\value{
table Information og data available for bootstrapping
}
\description{
Show stratification specified by control object, see (   ).
}
\examples{
\dontrun{check_for_missing_info_before_strata_aggregation(stom)}
}
