% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMdiet.R
\name{c.STOMdiet}
\alias{c.STOMdiet}
\title{Combine multiple STOMdiet objects}
\usage{
\method{c}{STOMdiet}(...)
}
\arguments{
\item{...}{STOMdiet objects to put together.}
}
\value{
Combined dataset of class STOMdiet.
}
\description{
Method to combine many datasets of the class STOMdiet
}
\details{
This function combines STOMdiet objects by \code{dplyr::bind_rows} using each of the
two components of the objects.
The method is useful for combining diet data from a number of predator that do not use the same raising procedures.
If the same stomach_id is present within different datasets the method will stop.
If a variable name is only present in some of the datasets, the variable will be
added to the remaining datasets (filled with NA) and a warning will be triggered.
The control attributes from the first STOMdiet will be used for the resulting combined STOMdiet
The R-Code is modified from the DATRAS R-package (Thanks to Casper Berg)
}
\examples{
\dontrun{ab<-c(a,b)}
}
