% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMobs.R
\name{c.STOMobs}
\alias{c.STOMobs}
\title{Combine multiple STOMobs objects}
\usage{
\method{c}{STOMobs}(...)
}
\arguments{
\item{...}{STOMobs objects to put together.}
}
\value{
Combined dataset of class STOMobs.
}
\description{
Method to combine many datasets
}
\details{
This function combines STOMobs objects by \code{dplyr::bind_rows} using each of the
two components of the objects.
The method is useful when reading large amounts of data in small chunks.
If the same stomach_id is present within different datasets the method will stop.
If a variable name is only present in some of the datasets, the variable will be
added to the remaining datasets (filled with NA) and a warning will be triggered.
The control attributes from the first STOMobs will be used for the resulting combined STOMobs
The R-Code is modified from the DATRAS R-package (Thanks to Casper Berg)
}
