% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_data.R
\name{change_data}
\alias{change_data}
\title{Make changes on  STOMobs from user options}
\usage{
change_data(
  stom,
  delete_vars,
  control_criteria = FALSE,
  pred_weight_multiplier = 1,
  pred_length_multiplier = 1,
  prey_weight_multiplier = 1,
  prey_length_multiplier = 1,
  insert_quarter = TRUE,
  correct_nstom = TRUE,
  correct_predl = TRUE,
  correct_preyl = TRUE,
  add_weight_method = FALSE,
  add_record_type = TRUE,
  reorganize_keys = FALSE
)
}
\arguments{
\item{stom}{Stomach data set of class STOMobs.}

\item{delete_vars}{vector of variable names to be deleted.}

\item{control_criteria}{dplyr::select sub-set of data from criteria given in control.
That is criteria given by the slots predators, years, stratum_areas and stratum. See \linkS4class{STOMcontrol}.}

\item{pred_weight_multiplier}{Multiplier for predator body weight, eg.g from kg to grammes.}

\item{pred_length_multiplier}{Multiplier for predator body length, from cm to mm.}

\item{prey_weight_multiplier}{Multiplier for prey body weight.}

\item{prey_length_multiplier}{Multiplier for prey body length.}

\item{insert_quarter}{Insert quarter of the year field from date.}

\item{correct_nstom}{Insert values for number of stomachs (total, with food, regurgitated, empty) where missing.}

\item{correct_predl}{Insert "missing" \code{control@mis_l} values for predator length range (pred_ll and pred_lu) from pred_l where missing. Insert  missing pred_l from mean of pred_ll and pred_lu.}

\item{correct_preyl}{Insert "missing" \code{control@mis_l} values for prey length (prey_l) and range of prey  (prey_ll and prey_lu) where missing.}

\item{add_weight_method}{Add code,if missing, for method of weighing and recording the weights of preys item (r=by record, p=by prey species and digestion stage).}

\item{add_record_type}{Add code, if missing, for sampling method for stomachs (SS=Single stomachs, PS=Pooled stomachs).}

\item{reorganize_keys}{Change internal keys (\code{sample_id}, \code{fish_id}). Do not change into TRUE!.}
}
\value{
STOMobs object.
}
\description{
Change stomach data
}
\examples{
\dontrun{read_exchange_data(stom_dir = 'NorthSea', exchange_file = 'cod_stomachs_2017.dat')}
}
