% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_stom.R
\name{check_for_missing_info_before_strata_aggregation}
\alias{check_for_missing_info_before_strata_aggregation}
\title{Check if data are ready for aggregation.}
\usage{
check_for_missing_info_before_strata_aggregation(stom)
}
\arguments{
\item{stom}{Stomach data set of class STOMobs.}
}
\value{
logical
}
\description{
Check if data are ready for aggregation.
}
\examples{
\dontrun{check_for_missing_info_before_strata_aggregation(stom)}
}
