% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unique_sample_id.R
\name{check_unique_diet_id}
\alias{check_unique_diet_id}
\title{Check uniqueness of diet_id (key)}
\usage{
check_unique_diet_id(diet)
}
\arguments{
\item{diet}{Object of class STOMdiet.}
}
\value{
data frame with potential issues.
}
\description{
Check uniqueness of diet_id (key)
}
\examples{
\dontrun{problems <- check_unique_diet_id(diet=a)}
}
