% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unique_sample_id.R
\name{check_unique_sample_id}
\alias{check_unique_sample_id}
\title{Check uniqueness of sample_id and fish_id}
\usage{
check_unique_sample_id(stom)
}
\arguments{
\item{stom}{Stomach data using internal variable names.}
}
\value{
data frame with potential issues.
}
\description{
Check that within one combination of sample_id and fish_id there is only one set of predator information.
}
\examples{
\dontrun{problems <- check_unique_sample_id(stom=a)}
}
