% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_years.R
\name{combine_years}
\alias{combine_years}
\title{Combine years of stomach data.}
\usage{
combine_years(stom, year_file, verbose = FALSE)
}
\arguments{
\item{stom}{Stomach data set of class STOMobs, from e.g. \link{read_exchange_data}.}

\item{year_file}{File Name for csv file including the variable year_old, year_new and predator.}

\item{verbose}{Logical for showing details}
}
\value{
Stomach data of class STOMobs with where the years are changed according to the specifications in file year_file..
}
\description{
Combine years of stomach data, e.g. assign stomach data from 2014 and 2015 to the same year, e.g. 2014
}
\examples{
\dontrun{combine_year(s,year_file="combine_year_10.csv")}
}
