% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMdiet.R
\name{diet_prey_to_other}
\alias{diet_prey_to_other}
\title{Transform preys into "other"}
\usage{
diet_prey_to_other(diet, preys_to_other)
}
\arguments{
\item{diet}{Diet data of class STOMdiet.}

\item{preys_to_other}{Prey or preys to be added to the "other" prey group.}
}
\description{
Transform preys into "other"
}
