% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMdiet.R
\name{diet_relative}
\alias{diet_relative}
\title{Calculate diet data as relative weight}
\usage{
diet_relative(diet)
}
\arguments{
\item{diet}{Diet data of class STOMdiet.}
}
\value{
Diet data of class STOMdiet.
}
\description{
Calculate diet data as relative weight
}
