% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMdiet.R
\name{diet_sum}
\alias{diet_sum}
\title{Sum diet data}
\usage{
diet_sum(diet)
}
\arguments{
\item{diet}{Diet data of class STOMdiet.}
}
\value{
Diet data of class STOMdiet summed for each combination of prey species and prey size.
}
\description{
Sum diet data
}
