% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_output.R
\name{do_detailed_output}
\alias{do_detailed_output}
\title{Show a subset from test criteria}
\usage{
do_detailed_output(
  stom,
  append = TRUE,
  to_screen = FALSE,
  label,
  digits = 1,
  rel_weight = FALSE,
  write_criteria = FALSE,
  use_criteria = TRUE,
  show_mis_length = TRUE,
  transpose = FALSE,
  by_sample_id = FALSE
)
}
\arguments{
\item{stom}{Stomach data set of class STOMobs.}

\item{append}{logical, append to output file?}

\item{to_screen}{logical, write results on the screen.}

\item{label}{text included in output.}

\item{digits}{number of significant digits in output.}

\item{rel_weight}{logical for relative or absolute weight.}

\item{write_criteria}{Write selection  criteria.}

\item{use_criteria}{Use the criteria for a subset stored in the control attribute is selecting data.}

\item{show_mis_length}{Include preys with missing length information.}

\item{transpose}{Transpose output table.}

\item{by_sample_id}{Output by sample id.}
}
\value{
test output.
}
\description{
Make detailed summary output from a STOMobs object using a subset of data defined in the control attributes.
}
\examples{
 \dontrun{do_detailed_output(stom)}
}
