% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_output.R
\name{do_detailed_output_diet}
\alias{do_detailed_output_diet}
\title{Show a subset from test criteria}
\usage{
do_detailed_output_diet(
  diet,
  append = TRUE,
  to_screen = FALSE,
  label,
  digits = 1,
  rel_weight = FALSE,
  write_criteria = FALSE,
  transpose = FALSE
)
}
\arguments{
\item{append}{logical, append to output file?}

\item{to_screen}{logical, write results on the screen.}

\item{label}{text included in output.}

\item{digits}{number of significant digits in output.}

\item{rel_weight}{logical for relative or absolute weight.}

\item{write_criteria}{logical, write criteria for sub-setting.}

\item{transpose}{Transpose output table.}

\item{stom}{data set of class STOMdiet.}
}
\value{
test output.
}
\description{
Make detailed summary output from a STOMdiet object using a subset of data defined in the control attributes.
}
\examples{
 \dontrun{do_detailed_output_diet(stom)}
}
