% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stom.control.R
\name{edit_control}
\alias{edit_control}
\title{Edit control attribute}
\usage{
edit_control(x, ...)
}
\arguments{
\item{x}{data of class STOMobs.}

\item{...}{One or more settings value, see XXXX}
}
\value{
Data set of class STOMobs with changed control attributes
}
\description{
Edit control attribute
}
