% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_stom.R
\name{from_to_species_diet}
\alias{from_to_species_diet}
\title{Change predator or prey names.}
\usage{
from_to_species_diet(
  d,
  pred_from_to = c("xcode", "xshort"),
  prey_from_to = c("xSMS_species", "xshort"),
  sp_info_file = file.path(config_dir, "species_info.csv"),
  refactor = TRUE
)
}
\arguments{
\item{d}{Diet data set of class STOMdiet.}

\item{pred_from_to}{Change predator name (variable pred_name) from an external lookup table using first element of pred_from_to as key and second element of pred_from_to as new name.}

\item{prey_from_to}{Change prey name Same as above, but for prey name.}

\item{sp_info_file}{CSV file with the fields used for \code{pred_from_to}  and \code{prey_from_to} and and additional field "number" used for arranging predator and prey names as factors.}

\item{refactor}{Logical for re-factor of predator and prey names in accordance with the sequence given by the "number" variable in the \code{sp_info_file}}
}
\value{
update data of class STOMdiet.
}
\description{
Change predator or prey names.
}
