% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stom.control.R
\name{get_control}
\alias{get_control}
\title{Get control attribute}
\usage{
get_control(x)
}
\arguments{
\item{x}{data of class STOMobs.}
}
\value{
Data of class STOMcontrol
}
\description{
Get control attribute
}
