% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stom.control.R
\name{get_option_logical}
\alias{get_option_logical}
\title{Get option value from user defined options}
\usage{
get_option_logical(opt)
}
\arguments{
\item{opt}{}
}
\value{
Value of the specified option.
}
\description{
Get option value from user defined options
}
\examples{
\dontrun{get_option_logical('split_horsemackerel')}
}
