% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMobs.R
\name{length.STOMobs}
\alias{length.STOMobs}
\title{Length of predator file}
\usage{
\method{length}{STOMobs}(x)
}
\arguments{
\item{x}{Object of class STOMobs}

\item{show_attributes}{Show the attributes for the objects.}
}
\description{
Length of predator file
}
