% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_output.R
\name{make_criteria_set}
\alias{make_criteria_set}
\title{Extract a subset from test criteria}
\usage{
make_criteria_set(stom)
}
\arguments{
\item{stom}{Stomach data set of class STOMobs.}
}
\value{
STOMobs object.
}
\description{
Extract a subset stomach contents data (of an object of class STOMobs) from criteria provided by control@detailed_tst_criteria.
This function is useful for tracking a subset of STOMobs though series of data manipulations.
}
\examples{
 \dontrun{make_criteria_set(stom)}
}
