% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_length_classess.R
\name{make_length_classess}
\alias{make_length_classess}
\title{Put size classes on predator from observed length.}
\usage{
make_length_classess(
  inp_dir = ".",
  inp_file = "length_classes_config.csv",
  max_l = 9999,
  write_output = TRUE,
  out_file = "length_classes.csv",
  minus_one = -1,
  same_no_all_data = FALSE
)
}
\arguments{
\item{inp_dir}{Input directory for \code{inn_file}.}

\item{inp_file}{File name with the configuration of length classes with the format as outlined above. The file must be a CSV file}

\item{max_l}{Upper length in the last "largest" length class.}

\item{write_output}{Flag for saving the length classes defined on \code{out_file}}

\item{out_file}{Output file with defined length classes.}

\item{minus_one}{Value to define upper length in length class, e.g. \code{minus_one}= -1 will provide length classes "050-059" and "060-69" for l 50, 60 and 70.}

\item{same_no_all_data}{Logical, should all size classes have the same numbering irrespective of year and quarter?}
}
\value{
defined length classes.
}
\description{
Define length classes for predator and preys.
}
\details{
This function makes length classes defined from a lower and an upper value of length (or any other size measurement)
The definition of length classes is given in an input file \code{inp_file},

see this table
\tabular{rrrrrr}{
\strong{l} \tab \strong{Species} \tab \strong{y1} \tab \strong{y2} \tab \strong{q1} \tab \strong{q2} \cr
50 \tab  ALL     \tab  1991 \tab  2013 \tab  1  \tab    4 \cr
60 \tab  ALL     \tab  1991 \tab  2013 \tab  1  \tab    4 \cr
70 \tab  ALL     \tab  1991 \tab  2013 \tab  1  \tab    4 \cr
b   80 \tab  ALL     \tab  1991 \tab  2013 \tab  1  \tab    4 \cr
}
where
l is the length used to define length classes, e.g. the first two lines will produce the sizes "050-059" and "060-69".
Species is the species name for the length classes. If Species=ALL, all the species will get the same lenngth classes
y1 and y2 are the year range for which the length  classes will be used.
q1 and q2 are the year range for which the length  classes will be used.

The defined length classes are made for all combinations of Species, y1 to y1, q1 to q2.
}
\examples{
\dontrun{ll<-make_length_classess(inp_dir=system.file('extdata', package = 'FishStomachs'),
                          inp_file='length_classes_config.csv'))}
}
