% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_stom.R
\name{make_template_strata_weighting}
\alias{make_template_strata_weighting}
\title{Make a template for weighting factors for population diet.}
\usage{
make_template_strata_weighting(
  stom,
  strata = c("sub_strata", "strata", "total")[1],
  write_CSV = FALSE
)
}
\arguments{
\item{stom}{Stomach data set of class STOMobs.}

\item{strata}{Level for stratification.  One of 'sub_strata', 'strata' or 'total'}

\item{write_CSV}{Logical for write of a CSV file in the}
}
\value{
logical
}
\description{
Make a template for weighting factors.
}
\examples{
\dontrun{make_template_strata_weighting(stom)}
}
