% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_stom.R
\name{plot.STOMdiet}
\alias{plot.STOMdiet}
\title{Plot diet data.}
\usage{
\method{plot}{STOMdiet}(
  d,
  show_plot = TRUE,
  cut_pred_size = c(1, 10),
  addTitle = FALSE,
  tAngle = 90,
  addNstom = FALSE,
  nstomAngle = 45,
  Ncol = 2,
  Nrow = NULL,
  Strip.position = c("top", "bottom", "left", "right"),
  Colours,
  otherCol = "grey",
  refac_prey = FALSE
)
}
\arguments{
\item{d}{Diet data set of class STOMdiet.}

\item{show_plot}{Show the resulting graphs on screen (or save the results for later processing)}

\item{cut_pred_size}{From to in substring of predator size}

\item{addTitle}{Add predator name on top of the plot.}

\item{tAngle}{Angle X-axis text.}

\item{addNstom}{Show number of stomachs sampled}

\item{nstomAngle}{Angle number of stomachs text}

\item{Ncol}{number of columns in facet plot}

\item{Nrow}{number of rows in facet plot}

\item{Strip.position}{strip.position: "top" | "bottom" | "left" | "right"}

\item{Colours}{vector of colours for preys.}

\item{otherCol}{Colour for "other prey"}

\item{refac_prey}{Reorder preys}
}
\value{
nothing (if show_plot=TRUE) or a list of plots.
}
\description{
Plot diet data.
}
