% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{plotboots}
\alias{plotboots}
\title{Plot bootstrapping diet data.}
\usage{
plotboots(
  b,
  show_plot = TRUE,
  cut_pred_size = c(1, 10),
  addTitle = FALSE,
  tAngle = 90,
  Colours = "red",
  maxbins = 50
)
}
\arguments{
\item{b}{Bootstrap diet data set of class STOMdiet.}

\item{show_plot}{Show the resulting graphs on screen (or save the results for later processing)}

\item{cut_pred_size}{From to in substring of predator size}

\item{addTitle}{Add predator name on top of the plot.}

\item{tAngle}{Angle X-axis text.}

\item{Colours}{Colours for frequncy.}

\item{maxbins}{maximum number of bins in plot.}
}
\value{
nothing (if show_plot=TRUE) or a list of plots.
}
\description{
Plot bootstrapping diet data.
}
