% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_stom.R
\name{plotdif}
\alias{plotdif}
\title{Plot difference between two diet data set.}
\usage{
plotdif(
  d1,
  d2,
  relative = TRUE,
  show_plot = TRUE,
  cut_pred_size = c(1, 10),
  cut_prey_size = c(1, 10),
  addTitle = FALSE,
  tAngle = 90,
  Colours = c("green", "red", "blue"),
  refac_prey = FALSE,
  maxDif = 5,
  size_NA = 4,
  byVar = c("year-quarter", "year", "quarter", "none")[1]
)
}
\arguments{
\item{d1}{Diet data set of class STOMdiet.}

\item{d2}{Diet data set of class STOMdiet.}

\item{relative}{Logical, relative weight difference is used if TRUE, else absolute weight difference}

\item{show_plot}{Show the resulting graphs on screen (or save the results for later processing)}

\item{cut_pred_size}{From to in substring of predator size}

\item{cut_prey_size}{From to in substring of prey size}

\item{addTitle}{Add predator name on top of the plot.}

\item{tAngle}{Angle X-axis text.}

\item{Colours}{Colours for positive and negative difference (d1-d2)}

\item{refac_prey}{Reorder preys}

\item{maxDif}{maximum difference to be shown}

\item{size_NA}{Size of missing observation}

\item{byVar}{Make individual plots by combinations of 'year-quarter','year' or 'quarter' or lump all data together ('none').}
}
\value{
nothing (if show_plot=TRUE) or a list of plots.
}
\description{
Plot difference between two diet data set.
}
