% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_stom.R
\name{plotSize}
\alias{plotSize}
\title{Plot diet data.}
\usage{
plotSize(
  d,
  show_plot = TRUE,
  cut_pred_size = c(1, 10),
  addTitle = FALSE,
  tAngle = 90,
  Colours,
  otherCol = "grey",
  refac_prey = FALSE,
  byVar = c("year-quarter", "year", "quarter", "none")[1]
)
}
\arguments{
\item{d}{Diet data set of class STOMdiet.}

\item{show_plot}{Show the resulting graphs on screen (or save the results for later processing)}

\item{cut_pred_size}{From to in substring of predator size}

\item{addTitle}{Add predator name on top of the plot.}

\item{tAngle}{Angle X-axis text.}

\item{Colours}{vector of colours for preys.}

\item{otherCol}{Colour for "other prey"}

\item{refac_prey}{Reorder preys}

\item{byVar}{Make individual plots by combinations of 'year-quarter','year' or'quarter' or lump all data together ('none').}
}
\value{
nothing (if show_plot=TRUE) or a list of plots.
}
\description{
Plot diet data.
}
