% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_sample_id_size.R
\name{pool_data}
\alias{pool_data}
\title{Pool stomach contents within sample and size}
\usage{
pool_data(stom)
}
\arguments{
\item{stom}{stomach contents data of class STOMobs.}
}
\value{
stomach contents data of class STOMobs with pooled stomach contents and number of stomachs, and updated predator information.
}
\description{
Individual stomach's data are pooled and predator information are updated for samples with the same \code{sample_id} and predator size class.
}
\details{
The stomach contents weight by prey and prey_size class are summed for for all samples within a sample_id and predator size class.
Number of stomachs are similarly added to include all stomachs.
In the pooled data, the variable \code{fish_id} becomes the value of a previously assigned predator size.

Size classes of both predators and preys must have been assigned ( \link{put_size_class_on_predator} and \link{put_size_class_on_prey} ) before aggregation can be made.
Technically, the \code{fish_id} is updated with the predator size \code{pred_size}
}
\examples{
\dontrun{aggregate_within_sample(ns)}
}
