% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_prey_weights_from_pooled_weight.R
\name{prey_w_from_pooled_weight}
\alias{prey_w_from_pooled_weight}
\title{Allocate weight to individual prey lengths from prey-pooled weights}
\usage{
prey_w_from_pooled_weight(
  stom,
  sel_preys = c("Clupea harengus"),
  sum_other_preys = TRUE,
  do_plots = FALSE,
  quantile_lw = c(0.01, 0.99)
)
}
\arguments{
\item{stom}{input Stomach contents data of class STOMobs.}

\item{sel_preys}{Vector of prey names for processing.}

\item{sum_other_preys}{Flag for transforming preys not included in sel_preys into prey="other food" with no length information}

\item{do_plots}{Flag for plotting derived length-weight relations and other plots.}

\item{quantile_lw}{Quantile of size range of prey length to be used for estimating length-weight relation}
}
\value{
Stomach data, class STOMobs.
}
\description{
Allocate weight to individual prey lengths from prey-pooled weights using length-weight relations by prey species
and digestion stage from records where prey weights are recoderde by the individual record.
This is done for the species included in the list of species (control@sel_species).
For other species, a total weight by species is calculated and length information is set to missing.
}
\examples{
\dontrun{individual_prey_weights_from_pooled_weight(stom=a,sel_preys=c("Clupea harengus"),
      sum_other_preys=TRUE,do_plots=FALSE)}
}
