% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stom.control.R
\name{print,STOMcontrol-method}
\alias{print,STOMcontrol-method}
\title{Print control object}
\usage{
\S4method{print}{STOMcontrol}(
  x,
  show = c("general", "data_used", "stratification", "calc_diet", "model_options", "ALK",
    "constants", "test_output", "detailed_test_output", "bootstrapping")
)
}
\arguments{
\item{x}{object of class STOMcontrol or STOMobs.}

\item{show}{Show values for selected group of options. Possible groups are 'general','data_used','stratification','calc_diet','model_options','ALK','constants','test_output' and 'detailed_test_output'.}
}
\description{
Print control object
}
