% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMdiet.R
\name{print.STOMdiet}
\alias{print.STOMdiet}
\title{Print overview of diet data}
\usage{
\method{print}{STOMdiet}(x)
}
\arguments{
\item{x}{Diet data of class STOMdiet.}
}
\description{
Print overview of diet data
}
