% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMobs.R
\name{print.STOMobs}
\alias{print.STOMobs}
\title{Print STOMobs}
\usage{
\method{print}{STOMobs}(x, show_attributes = TRUE)
}
\arguments{
\item{x}{Object of class STOMobs}

\item{show_attributes}{Show the attributes for for the objects.}
}
\description{
Print STOMobs
}
