% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_length_classes_on.R
\name{put_size_class_on_predator}
\alias{put_size_class_on_predator}
\title{Put size classes on predator from observed length.}
\usage{
put_size_class_on_predator(stom, len_classes)
}
\arguments{
\item{stom}{stomach contents data of class STOMobs.}

\item{len_classes}{Data frame with size class definitions. See \code{make_length_classes}. Size class will be added just for the species included in \code{len_classes}. If \code{length_classes} include sprecies='ALL' all species will get size class.}
}
\value{
stomach contents data of class STOMobs wit added size class for predators.
}
\description{
Put size classes on predators
}
\details{
This function adds a size group to each predator based on the predator length (pred_l) if such exist,
or based on the midpoint of the size range (pred_ll - pred_lu). Both approaches use an input definition of size groups,
typically made by the function \code{make_length_classes}.
The mean length of the predator (pred_l_mean) is added either copied from the variable pred_l or from the midpoint of the size range (pred_ll - pred_lu)
The size clases are species independent if \code{len_classes} includes Species==ALL. .
If length classes is given by predator in \code{len_classes}, species specific length classes will be added.
}
