% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_length_classes_on.R
\name{put_size_class_on_prey}
\alias{put_size_class_on_prey}
\title{Put size classes on preys from observed length.}
\usage{
put_size_class_on_prey(stom, len_classes, too_small_to_other = TRUE)
}
\arguments{
\item{stom}{stomach contents data of class STOMobs.}

\item{len_classes}{Data frame with size class definitions. See \code{make_length_classes}. Size class will be added just for the species included in \code{len_classes}. If \code{length_classes} include sprecies='ALL' all species will get size class.}

\item{too_small_to_other}{Convert small prey items into other food. Preys smaller than defined by control@min_prey_length are renamed into other food with unknown size if \code{too_small_to_other} is TRUE.}
}
\value{
stomach contents data of class STOMobs with added size class for predators.
}
\description{
Put size classes on preys
}
\details{
This function adds a size group to each prey item based on the prey length (prey_l) if such exist,
or based on the midpoint of the size range (prey_ll - prey_lu). Both approaches use an input definition of size groups,
typically made by the function \code{make_length_classes}.
The mean length of the predator (XXXXX) is added either copied from the variable prey_l or from the midpoint of the size range (prey_ll - prey_lu)
The size clases are species independent if \code{len_classes} includes Species=ALL. .
If length classes is given by predator in \code{len_classes}, species specific length classes will be added.
}
