% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exchange_data.R
\name{read_exchange_data}
\alias{read_exchange_data}
\title{Read stomach contents data on exchange format}
\usage{
read_exchange_data(
  control,
  delete_errors = FALSE,
  allow_alias_names = FALSE,
  keep_just_mandatory_fields = FALSE,
  fileEncoding = ""
)
}
\arguments{
\item{control}{Options \link{STOMcontrol-class}) for dplyr::selecting and collating stomach data.}

\item{delete_errors}{Logical for deleting records with errors.
if FALSE, the function will stop at the first error found,
if TRUE records with errors will be deleted, but the function continues if possible.}

\item{allow_alias_names}{Logical for using alias field names as specified in the stomach_format}

\item{keep_just_mandatory_fields}{Logical for just keeping mandatory field names as specified in the stomach_format}

\item{fileEncoding}{character string: if non-empty declares the encoding used on a file so the character data can be re-encoded.}
}
\value{
Stomach data of class STOMobs.
}
\description{
Read data from exchange data
}
\examples{
\dontrun{read_exchange_data(control, fileEncoding = "latin1")}
}
