% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redistribution.R
\name{redist_unidentified_prey_lengths}
\alias{redist_unidentified_prey_lengths}
\title{Redistribute unidentified prey lengths}
\usage{
redist_unidentified_prey_lengths(
  s,
  dist_time = stratum_time,
  dist_area = stratum_area,
  dist_pred_size = pred_size,
  selected_pred,
  remains_to_other = TRUE,
  others_to_other = TRUE
)
}
\arguments{
\item{s}{Stomach contents data of class STOMobs.}

\item{dist_time}{Stratum used for time in constructing proportion of species within a prey group of not fully identified prey items. The variable should be made from variable names in s, or constants.}

\item{dist_area}{Stratum used for area in constructing proportion of species within a prey group of not fully identified prey items. The variable should be made from variable names in s, or constants.}

\item{dist_pred_size}{Stratum used for predator size in constructing proportion of species within a prey group of not fully identified prey items. The variable should be made from variable names in s, or constants.}

\item{selected_pred}{Predators for which redistribution is done. All predators included in s is selected if selected_pred value is missing.}

\item{remains_to_other}{Convert not fully identified prey species to prey 'other' where no species allocation key cannot be found.}

\item{others_to_other}{Convert not species not included in \code{selected_pred} to prey 'other'.}
}
\value{
s Stomach contents data of class STOMobs.
}
\description{
Redistribute preys with no length information
}
\details{
Redistribute missing length observations using the observed length distribution of the species '
from the same predator and predator size class within the given area and time strata. Only prey species '
specified by control@sel_preys are redistributed.'
}
