% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stom.control.R
\docType{class}
\name{STOMcontrol-class}
\alias{STOMcontrol-class}
\title{An s4 class for Control object for stomach compilation.}
\value{
object of class STOMcontrol
}
\description{
An s4 class for Control object for stomach compilation.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Name of control object.}

\item{\code{stomach_format}}{\code{expression} (R code) for file name for CSV file defining stomach contents data.
Default: expression(system.file("stomach_format.csv", package = "FishStomachs"))
For more details on exchange format see the vignette: \code{vignette("Database_format", package = "FishStomachs")}}

\item{\code{stomach_dir}}{characters. Directory name with stomach data on exchange format.}

\item{\code{dataSets}}{File name or file names for data on exchange format.}

\item{\code{years}}{Year or years of data to be used.}

\item{\code{quarters}}{Quarters of the year ( or similar time split) to be used.}

\item{\code{predators}}{Predator name or names.}

\item{\code{strata_area_exp}}{\code{expression} for defining strata area.}

\item{\code{strata_sub_area_exp}}{\code{expression} for defining sub_area.}

\item{\code{strata_time_exp}}{\code{expression} for defining temporal strata.}

\item{\code{strata_year_back}}{\code{expression} for getting year information out of diet data, e.g. \code{expression(as.numeric(substr(stratum_time,1,4)))}.}

\item{\code{strata_quarter_back}}{\code{expression} for getting quarter information out of diet data.}

\item{\code{sel_preys}}{Selected prey names for diet calculation.}

\item{\code{min_prey_length}}{Minimum prey length (or included in other food) \code{integer}.}

\item{\code{calc_strata}}{Specification of strata}

\item{\code{weighting_factor}}{Weighting factor for weighted mean}

\item{\code{year_quarter_strata}}{vector.}

\item{\code{weighting_by_nstom}}{logical.}

\item{\code{combine_starta_from_relative_stomach_content}}{logical.}

\item{\code{model_options}}{list. Options for producing input values to multi species models as SMS.
\preformatted{The list must as a minimum include the names :
 "delete_small" logical, delete input with relative stomach contents weight lower than "min_value".
                if FALSE values lower than "min_value" are increased to "min_value".
 "min_value", minimum relative stomach contents weight (see above)
 "insert_mid_value" logical, insert dummy value ("mid_value") where observations for a prey  within a size range are missing
 "mid_value" value used for missing mid-values.
 "insert_tails", logical for insertion of dummy values("tail_value") for the prey size class which are lower and higher than the observed size range for a prey
 "tail_value",  value used for missing tail-values.
 }}

\item{\code{max_l}}{integer. Number used for the maximum length, e.g. 9998L}

\item{\code{mis_l}}{integer. Number used for missing length, e.g. 9999L}

\item{\code{mis_digest}}{integer. Number used for missing digestion stage e.g. 9L}

\item{\code{other}}{character. Name for the "Other food", which is prey items not considered as individual prey species for further analysis, e.g. "other".}

\item{\code{mis_size_class}}{integer. Number for size class where the size is unknown, e.g. 0L}

\item{\code{ALK_options}}{list, options used for creating Age Length Keys}

\item{\code{do_test_output}}{logical.}

\item{\code{do_test_output_file}}{logical.}

\item{\code{tst_file}}{character.}

\item{\code{detailed_tst_output}}{logical.}

\item{\code{detailed_tst_file}}{character.}

\item{\code{detailed_tst_criteria}}{list.}

\item{\code{boots_id}}{list of \code{expression} for defining sampling unit (haul) from STOMobs variables.}

\item{\code{boots_strata}}{list \code{expression} for defining strata (haul).}
}}

\examples{
\dontrun{control<-new("STOMcontrol",
      name='Baltic cod',
      stomach_dir=file.path(stom_data_dir,'baltic'),
      dataSets= c("balt_stom_new_formatNO_new_var.csv"))
      }
}
