% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_stom.R
\name{strata_aggregate}
\alias{strata_aggregate}
\title{Aggregate stoamch contents within strata (time, area and sub-area)}
\usage{
strata_aggregate(stom)
}
\arguments{
\item{stom}{stomach contents data of class STOMobs.}
}
\value{
stomach contents data of class STOMstrata.
}
\description{
Aggregate stomach contents within strata (time, area and sub-area)
}
\details{
Stomach contents data are aggregated (sum) within each combination of stratum_time, stratum_area and stratum_sub_area.

The number of stomachs (n_tot) is dplyr::summarise within the strata.
}
\examples{
\dontrun{b<-strata_aggegate(stom)}
}
