% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMdiet.R
\name{subset.STOMdiet}
\alias{subset.STOMdiet}
\title{Sub-setting a STOMdiet object.}
\usage{
\method{subset}{STOMdiet}(x, ..., na_rm = TRUE)
}
\arguments{
\item{x}{Take subset of this dataset.}

\item{...}{One or more subset criteria.}

\item{na_rm}{Delete missing values.}
}
\value{
The reduced dataset of class STOMdiet.
}
\description{
General method for sub-setting a STOMdiet object.
}
\details{
A STOMdiet object contains two kinds of information:
\enumerate{
\item Between predator information (contained in component PRED)
\item Within prey information (contained in component PREY)
}
This method can subset over both kinds of information - at once.
The functionality is achieved by matching each of the subset critera
to the appropriate data component. When a match is found the subset is
performed on the component and the subsetting is continued to the next
subset criteria. The number of subset critera can be arbitrary and is
given through the \code{...} argument.
Code is copied from the DATRAS package
}
