% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMobs.R
\name{subset.STOMobs}
\alias{subset.STOMobs}
\title{Sub-setting a STOMobs object.}
\usage{
\method{subset}{STOMobs}(x, ..., na_rm = TRUE)
}
\arguments{
\item{x}{Take subset of this dataset.}

\item{...}{One or more subset criteria.}

\item{na_rm}{Discard missing values in subset criterion?}
}
\value{
subset of dataset of class STOMobs.
}
\description{
General method for sub-setting a STOMobs object.
}
\details{
A STOMobs object contains two kinds of information:
\enumerate{
\item Between predator information (contained in component PRED)
\item Within prey information (contained in component PREY)
}
This method can subset over both kinds of information - at once.
The functionality is achieved by matching each of the subset criteria
to the appropriate data component. When a match is found the subset is
performed on the component and the subsetting is continued to the next
subset criteria. The number of subset criteria can be arbitrary and is
given through the \code{...} argument.
Code is copied from the DATRAS package
}
\examples{
\dontrun{subset(ns,pred_name=='Cod')}
}
