% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMdiet.R
\name{summary.STOMdiet}
\alias{summary.STOMdiet}
\title{Summary of diet}
\usage{
\method{summary}{STOMdiet}(diet, level = 1, digits = 1, drop.unused.levels = FALSE)
}
\arguments{
\item{diet}{Diet data of class STOMdiet.}

\item{level}{Detail level.}

\item{digits}{Number of digits in output.}

\item{drop.unused.levels}{Drop unused factor levels.}
}
\description{
Summary of diet
}
