% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMobs.R
\name{summary.STOMobs}
\alias{summary.STOMobs}
\title{Summary of number of stomachs}
\usage{
\method{summary}{STOMobs}(x, level = 1)
}
\arguments{
\item{x}{Diet data of class STOMdiet.}

\item{level}{Level of details.}
}
\description{
Summary of number of stomachs
}
