% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_stom.R
\name{table_strata_prey_preyl}
\alias{table_strata_prey_preyl}
\title{Table stomach contents within by strata (time, area and sub-area)}
\usage{
table_strata_prey_preyl(s, use_criteria = FALSE, do_show = TRUE, digits = 1)
}
\arguments{
\item{s}{stomach contents data of class STOMstrata (produced by strata_aggegate)}

\item{use_criteria}{Use criteria in control to subset data}

\item{do_show}{Show the tables on the screen.}

\item{digits}{Number of digits in output.}
}
\value{
Tables of stomach contents.
}
\description{
Table stomach contents within by strata (time, area and sub-area). Tables are returned or shown on screen.
}
\examples{
\dontrun{b<-strata_aggegate(stom)}
}
