% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STOMobs.R
\name{update.STOMobs}
\alias{update.STOMobs}
\title{Update attributes with added variables}
\usage{
\method{update}{STOMobs}(stom)
}
\arguments{
\item{stom}{Stomach data set of class STOMobs.}
}
\value{
Stomach data set of class STOMobs with updated attributes.
}
\description{
Method to update variables names in the attributes for STOMobs.
The attributes with variable names are updated for the PRED and PREY data frame.
}
\examples{
\dontrun{stom<-update(stom); stom}
}
