% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stom.control.R
\name{validSTOMcontrol}
\alias{validSTOMcontrol}
\title{Validation of control object used for stomach data compilation.}
\usage{
validSTOMcontrol(object)
}
\arguments{
\item{object}{Object of class STOMcontrol}
}
\description{
Validation of control object used for stomach data compilation.
}
