% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{add_missing_boots}
\alias{add_missing_boots}
\title{Add missing prey and prey size combination for bootstrap replicates}
\usage{
add_missing_boots(bt, mis_value = 0)
}
\arguments{
\item{bt}{Bootstrapped diet data.}

\item{mis_value}{Value for missing prey weights.}
}
\value{
Bootstrapped diet data with inserted value for missing prey and prey size combinations
}
\description{
Add missing prey and prey size combination for bootstrap replicates
}
\examples{
\dontrun{s<-add_missing(boots(s,value=NA))}
}
