% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootsMean}
\alias{bootsMean}
\title{Statistics from bootstrap replicates.}
\usage{
bootsMean(
  b,
  pointEst,
  by_prey_size = FALSE,
  n_rep = NA,
  do_Diri = FALSE,
  minPreyProportion = 0,
  Diri_min = 0.001,
  excl_diri = NA,
  Diri_max = 0.99,
  verbose = FALSE
)
}
\arguments{
\item{b}{Bootstrap diet data set. List of elements of class STOMdiet.}

\item{pointEst}{Diet data estimated without bootstrapping (point estimates of prey weights)}

\item{by_prey_size}{Logical for calculating mean and variance by prey size. FALSE provides statistics by prey species.}

\item{n_rep}{Selected replicates used. If n_rep is a number, 1:n_rep are used, If n_rep is a vector the provided replica numbers are used in the statistics. All replicates are uesed if n_rep is NA (default).}

\item{do_Diri}{Logical for performing statistics assuming the replicates are Dirichlet distributed.}

\item{minPreyProportion}{Lower level for average prey proportion to be kept as a named prey. If lower, the prey weight is allocated to "other" prey.}

\item{Diri_min}{Value to be used in estimating the Dirichelt parameter in case of a zero proportion.}

\item{excl_diri}{Combinations (year, quarter, pred_name, pred_size) where Dirichlet parameters are not estimated.}

\item{Diri_max}{Maximum average proportion for one food item  allowed for estimation of of Dirichlet parameters.}

\item{verbose}{Show progress.}
}
\value{
Data set with mean and variance of diets and fit to Dirichlet distribution. Output includes
\tabular{ll}{
\strong{Variable} \tab \strong{Contents} \cr
nboots    \tab Number of bootstrap replicates.\cr
year      \tab Year.\cr
quarter   \tab Quarter of the year.\cr
pred_name \tab Predator name.\cr
pred_size \tab Predator size.\cr
prey      \tab Prey, prey name and size, if used.\cr
prey_name \tab Prey name.\cr
prey_size \tab Prey size.\cr
n_prey_sp \tab Number of "preys" observed.\cr
param     \tab Estimated parameters assuming a Dirichlet distribution of bootstrap replicates.\cr
phi       \tab Estimated precision parameter (sum of param from the Dirichlet distribution).\cr
mu        \tab Estimated mean value from the Dirichlet distribution.\cr
mean_w    \tab Mean value of prey weights (proportions) from the bootstrap replicates.\cr
sd_w      \tab Standard deviation of prey weights (proportions).\cr
prey_w    \tab Prey weight (proportions) estimated without bootstrapping. \cr
strata    \tab Strata used for bootstrap.\cr
n_stom    \tab Number of stomachs available before bootstrap ( from the \code{pointEst} data).\cr
n_sample  \tab Number of samples available before bootstrap.\cr
}
}
\description{
Statistics from bootstrap replicates.
}
