% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootstrap_data}
\alias{bootstrap_data}
\title{Check if data are ready for aggregation.}
\usage{
bootstrap_data(s, nfac = 1, replace = TRUE, seed = 0)
}
\arguments{
\item{s}{Stomach data set of class STOMobs.}

\item{nfac}{Number of bootstrap samples relative to original number of samples.}

\item{replace}{Logical for bootstrapping with replacement.}

\item{seed}{Random seed value.}
}
\value{
s Bootstrapped stomach data set of class STOMobs.
}
\description{
Make bootstrap data from stratification specified by control object, see (   ).
}
\examples{
\dontrun{check_for_missing_info_before_strata_aggregation(stom)}
}
