% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{bootstrap_show}
\alias{bootstrap_show}
\title{Show bootstrap stratification.}
\usage{
bootstrap_show(
  s,
  show = c("strata", "sample")[1],
  vari = c("stomach", "sample")[1]
)
}
\arguments{
\item{s}{Stomach data set of class STOMobs.}

\item{show}{Show either number of stomachs per strata or number of stomachs per samples available for bootstrapping.}

\item{vari}{Variable, number of stomachs or number of samples}
}
\value{
table Information on data available for bootstrapping
}
\description{
Show stratification specified by control object, see (   ).
}
\examples{
\dontrun{bootstrap_show(s)}
}
