% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{model_output}
\alias{model_output}
\title{Make model output}
\usage{
model_output(
  diet,
  length_classes_file,
  sp_info_file,
  intoSMSformat = FALSE,
  pred_is_coded = FALSE,
  prey_is_coded = FALSE
)
}
\arguments{
\item{diet}{Object of class STOMdiet.}

\item{length_classes_file}{bla}

\item{sp_info_file}{bla}

\item{intoSMSformat}{bla}

\item{pred_is_coded}{Logical; predator name is already on a short form (e.g. FAO 3 letter code)}

\item{prey_is_coded}{Logical; prey name is already on a short form (e.g. FAO 3 letter code)}
}
\value{
Stomach contents data of class STOMobs.
}
\description{
Model output for the SMS model are produced from a STOMdiet object.
}
\examples{
\dontrun{x=2}

}
