% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping.R
\name{plotboots}
\alias{plotboots}
\title{Plot bootstrapping diet data.}
\usage{
plotboots(
  b,
  show_plot = TRUE,
  freq = TRUE,
  cut_pred_size = c(1, 10),
  addTitle = FALSE,
  tAngle = 90,
  Colours = c("#C10534", "green", "blue", "pink", "skyblue"),
  ltys = c("dotted", "dotdash", "longdash"),
  maxbins = 75,
  statistics,
  inclData,
  add_mean_diri = TRUE,
  add_mean_boots = FALSE,
  add_no_boots = FALSE
)
}
\arguments{
\item{b}{List of Bootstrap diet data of class STOMdiet.}

\item{show_plot}{Show the resulting graphs on screen (or save the results for later processing)}

\item{freq}{Logical, show frequency (or density).}

\item{cut_pred_size}{From to in substring of predator size}

\item{addTitle}{Add predator name on top of the plot.}

\item{tAngle}{Angle X-axis text.}

\item{Colours}{Colours for frequency.}

\item{maxbins}{maximum number of bins in plot.}

\item{statistics}{Data set (estimated by call to bootsMean) including statistics to be shown on the plot.}

\item{inclData}{Date set with combinations of stratum_time, pred_name and pred_size to be included in the plot.}

\item{add_mean_diri}{logical, plot diet proportions estimated from Dirichlet parameters (Colours\link{3}).}

\item{add_mean_boots}{logcal, plot diet proportions estimated from the mean of bootstrapping replicates (Colours\link{4}).}

\item{add_no_boots}{logical, plot diet proportions without bootstrapping. (Colours\link{5}).}
}
\value{
nothing (if show_plot=TRUE) or a list of plots.
}
\description{
Plot bootstrapping diet data.
}
