% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stom.control.R
\docType{class}
\name{STOMcontrol-class}
\alias{STOMcontrol-class}
\title{An s4 class for Control object for stomach compilation}
\value{
a control object
}
\description{
An s4 class for Control object for stomach compilation
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Name of control object.}

\item{\code{stomach_format}}{expression. \code{expression} to get file name for file defining stomach contents data. Default: expression(system.file("stomach_format.csv", package = "FishStomachs"))}

\item{\code{stomach_dir}}{characters. Directory name for directory with stomach data on exchange format.}

\item{\code{dataSets}}{vector of characters. File name or file names for data on exchange format.}

\item{\code{years}}{vector. Year of data to be used.}

\item{\code{quarters}}{vector. Quarters of the year ( or similar time split) to be used.}

\item{\code{strata_area_exp}}{expression for defining strata area.}

\item{\code{strata_sub_area_exp}}{expression for defining sub_area.}

\item{\code{strata_time_exp}}{expression for defining temporal strata.}

\item{\code{strata_year_back}}{"expression" (R-code) for getting year information out of diet data, e.g. \code{expression(as.numeric(substr(stratum_time,1,4)))}.}

\item{\code{strata_quarter_back}}{"expression" (R-code) for getting quarter information out of diet data.}

\item{\code{sel_preys}}{vector.}

\item{\code{year_quarter_strata}}{vector.}

\item{\code{weighting_by_nstom}}{logical.}

\item{\code{combine_starta_from_relative_stomach_content}}{logical.}

\item{\code{min_prey_length}}{integer.}

\item{\code{max_l}}{integer. Number used for the maximum length, e.g. 9998L}

\item{\code{mis_l}}{integer. Number used for missing length, e.g. 9999L}

\item{\code{other}}{character. Name for the "Other food", which is prey items not considered as individual prey species for further analysis, e.g. "other".}

\item{\code{mis_size_class}}{integer. Number for size class where the size is unknown, e.g. 0L}

\item{\code{model_options}}{list. Options for producing input values to multi species models as SMS.}

\item{\code{do_test_output}}{logical.}

\item{\code{do_test_output_file}}{logical.}

\item{\code{tst_file}}{character.}

\item{\code{detailed_tst_output}}{logical.}

\item{\code{detailed_tst_file}}{character.}

\item{\code{detailed_tst_criteria}}{list.}
}}

\examples{
\dontrun{control<-new("STOMcontrol",
      name='Baltic cod',
      stomach_dir=file.path(stom_data_dir,'baltic'),
      dataSets= c("balt_stom_new_formatNO_new_var.csv"))
      }
}
