% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exchange_data.R
\name{write_exchange_data}
\alias{write_exchange_data}
\title{Write stomach contents data of class STOMobs on the exchange format.}
\usage{
write_exchange_data(stom, exchange_file = NA)
}
\arguments{
\item{stom}{Input stomach data set for writing}

\item{exchange_file}{File name for output data on exchange format.}
}
\value{
Returns the input stom invisibly.
}
\description{
Write data on exchange data
}
\examples{
\dontrun{write_exchange_data(tst, exchange_file = 'cod_stomachs_2017.csv')}
}
